finish 2B

+++ CS142-S25-Problem-Solving-HW2/A-Ski/main.cpp
@@ -4,9 +4,9 @@
 
 struct location
 {
-  int height;
-  int i;
-  int j;
+  unsigned int height;
+  unsigned int i;
+  unsigned int j;
 };
 
 int main()
@@ -14,11 +14,11 @@ int main()
   std::ios_base::sync_with_stdio(false);
   std::cin.tie(0);
 
-  int rows;
-  int columns;
+  unsigned int rows;
+  unsigned int columns;
   std::vector<location> locations;
 
-  int longest_route = 1;
+  unsigned int longest_route = 1;
 
   // define 2D movement
   const int direction_i[4] = {-1, 1, 0, 0};
@@ -27,20 +27,20 @@ int main()
   std::cin >> rows;
   std::cin >> columns;
 
-  std::vector<std::vector<int>> heights_input(rows, std::vector<int>(columns));
-  std::vector<std::vector<int>> dp(rows, std::vector<int>(columns, 1));
+  std::vector<std::vector<unsigned int>> heights_input(rows, std::vector<unsigned int>(columns));
+  std::vector<std::vector<unsigned int>> dp(rows, std::vector<unsigned int>(columns, 1));
 
-  for ( int i = 0; i < rows; ++i )
+  for ( unsigned int i = 0; i < rows; ++i )
   {
-    for ( int j = 0; j < columns; ++j )
+    for ( unsigned int j = 0; j < columns; ++j )
     {
       std::cin >> heights_input[i][j];
     }
   }
 
-  for ( int i = 0; i < rows; ++i )
+  for ( unsigned int i = 0; i < rows; ++i )
   {
-    for ( int j = 0; j < columns; ++j )
+    for ( unsigned int j = 0; j < columns; ++j )
     {
       locations.push_back({heights_input[i][j], i, j});
     }
@@ -65,30 +65,30 @@ int main()
 
   for ( const auto& location : locations )
   {
-    int current_height = location.height;
-    int i = location.i;
-    int j = location.j;
+    unsigned int current_height = location.height;
+    unsigned int i = location.i;
+    unsigned int j = location.j;
 
-    int current_dp_height = dp[i][j];
+    unsigned int current_dp_height = dp[i][j];
 
 
 
     // check neighbors
-    for ( int directions = 0; directions < 4; ++directions )
+    for ( unsigned int directions = 0; directions < 4; ++directions )
     {
-      int new_i = i + direction_i[directions];
-      int new_j = j + direction_j[directions];
+      int new_i = static_cast<int>(i) + direction_i[directions];
+      int new_j = static_cast<int>(j) + direction_j[directions];
 
 
 
       // check boundaries and height decrease
-      if ( new_i >= 0 && new_i < rows && new_j >= 0 && new_j < columns && heights_input[new_i][new_j] < current_height )
+      if ( static_cast<int>(new_i) >= 0 && static_cast<unsigned int>(new_i) < rows && static_cast<int>(new_j) >= 0 && static_cast<unsigned int>(new_j) < columns && heights_input[static_cast<unsigned int>(new_i)][static_cast<unsigned int>(new_j)] < current_height )
       {
-        dp[new_i][new_j] = std::max(dp[new_i][new_j], current_dp_height + 1);
+        dp[static_cast<unsigned int>(new_i)][static_cast<unsigned int>(new_j)] = std::max(dp[static_cast<unsigned int>(new_i)][static_cast<unsigned int>(new_j)], current_dp_height + 1);
 
 
 
-        longest_route = std::max(longest_route, dp[new_i][new_j]);
+        longest_route = std::max(longest_route, dp[static_cast<unsigned int>(new_i)][static_cast<unsigned int>(new_j)]);
       }
     }
   }
